feat: authorization implemented

+++ src/services/notes/controller/note-controller.js
@@ -2,15 +2,17 @@ import { nanoid } from 'nanoid';
 import NoteRepositories from '../repositories/note-repositories.js';
 import { InvariantError, NotFoundError } from '../../../exceptions/index.js';
 import response from '../../../utils/response.js';
+import AuthorizationError from '../../../exceptions/authorization-error.js';
 
 export const createNote = async (req, res, next) => {
   const { title, body, tags } = req.validated;
+  const { id: owner } = req.user;
   const note = await NoteRepositories.createNote({
     title,
     body,
     tags,
+    owner,
   });
-
   if (!note) {
     return next(new InvariantError('Catatan gagal ditambahkan'));
   }
@@ -18,47 +20,62 @@ export const createNote = async (req, res, next) => {
   return response(res, 201, 'Catatan berhasil ditambahkan', note);
 };
 
-export const getNotes = async (req, res, next) => {
-  const notes = await NoteRepositories.getNotes();
+export const getNotes = async (req, res) => {
+  const { id: owner } = req.user;
+  const notes = await NoteRepositories.getNotes(owner);
   return response(res, 200, 'Catatan sukses ditampilkan', notes);
 };
 
 export const getNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(
+      new AuthorizationError('Anda tidak berhak mengakses resource ini'),
+    );
+  }
   const note = await NoteRepositories.getNoteById(id);
-
   if (!note) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
-
   return response(res, 200, 'Catatan sukses ditampilkan', note);
 };
 
 export const updateNote = async (req, res, next) => {
   const { id } = req.params;
   const { title, body, tags } = req.validated;
-
-  const note = await NoteRepositories.updateNote({
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(
+      new AuthorizationError('Anda tidak berhak mengakses resource ini'),
+    );
+  }
+  const note = await NoteRepositories.editNote({
     id,
     title,
     body,
     tags,
   });
-
   if (!note) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
-
   return response(res, 200, 'Catatan berhasil diperbarui', note);
 };
 
 export const deleteNote = async (req, res, next) => {
   const { id } = req.params;
-  const deleteNote = await NoteRepositories.deleteNote(id);
-
-  if (!deleteNote) {
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(
+      new AuthorizationError('Anda tidak berhak mengakses resource ini'),
+    );
+  }
+  const deletedNote = await NoteRepositories.deleteNote(id);
+  if (!deletedNote) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
-
-  return response(res, 200, 'Catatan berhasil dihapus', deleteNote);
+  return response(res, 200, 'Catatan berhasil dihapus', deletedNote);
 };
